Split github package into several files
